Move BaseConfigurator from isntrumentation to SDK
This allows SDK to not depend on instrumentation package anymore.
