+++ opentelemetry-sdk/setup.cfg
@@ -44,7 +44,6 @@ include_package_data = True
 install_requires =
     opentelemetry-api == 1.6.0
     opentelemetry-semantic-conventions == 0.25b0
-    opentelemetry-instrumentation == 0.25b0
 
 [options.packages.find]
 where = src
